add code to update and maintain the following link tables:
- soil
- sun
- water

By renaming tables to fit a uniform pattern, it will be easy to maintain
linking tables in a generic way.

We should rethink the "_requirements" addition to the table names, as
well as renaming things like "attracts_wildlife" to just "wildlife" and
then have a plant_wildlife linking table.  Similar changes should be
made to flower_color, foliage_color, and any other attributes of a plant
where the plant can have more than one...

If we settle on a standard, then adding linking tables, and managing
them becomes pretty easy.

I wasn't too sure how far to go, so I'm leaving this in as is for now,
and we can chat on the phone about how they work, and the database
changes needed to support other linking tables easily.

I don't want Myrna to come back to a completely re-written
application...